纳秒精度的 log 时间戳

+++ modules/runtime/include/misc/log/log_base.hpp
@@ -37,7 +37,7 @@ protected:
     friend struct LogPattern;
     bool flush = false;
     LogLevel level;
-    uint32_t timestamp;
+    int64_t timestamp;
     uint64_t thread_id;
     const char8_t* thread_name;
     struct Logger* logger;
